Track time better

+++ mcmtestOpenTK/mcmtestOpenTK/Client/GlobalHandler/MainGame_Defs.cs
@@ -146,6 +146,6 @@ namespace mcmtestOpenTK.Client.GlobalHandler
         /// <summary>
         /// The current global time.
         /// </summary>
-        public static long GlobalTickTime;
+        public static double GlobalTickTime;
     }
 }
